Fix Post Preview Strip HTML Bug
Fixes #175.

If a post was truncated in the middle of an HTML tag, the HTML couldn't
be properly stripped and would show up on the page. This commit fixes
that.

The post preview feature uses 3 filters: markdownify, strip_html, and
truncatewords. We need to do truncatewords last so we don't truncate in
the middle of a tag.